+++ app/testblog/forms.py
@@ -1,6 +1,6 @@
 from trionyx import forms
 
-from app.testblog.models import Tag, Post
+from app.testblog.models import Tag, Post, Category
 
 
 @forms.register(code='new', default_edit=True)
@@ -14,6 +14,9 @@ class TagForm(forms.ModelForm):
 
 @forms.register(default_edit=True)
 class PostForm(forms.ModelForm):
+    category = forms.ModelAjaxChoiceField(Category.objects.all())
+    tags = forms.ModelAjaxMultipleChoiceField(Tag.objects.all(), required=False)
+
     class Meta:
         model = Post
         fields = ['title', 'content', 'publish_date', 'category', 'status', 'price', 'tags']